# Documentação/Collections com os endpoints/V1_swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Api AlterData",
    "description": "Documentação da API Alterdata",
    "contact": {
      "name": "Eglecio Alexandre Pereira",
      "url": "https://github.com/eglecio/AlterData",
      "email": "eglecio@gmail.com"
    },
    "version": "v1"
  },
  "paths": {
    "/Cliente": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "summary": "Adiciona um novo cliente no sistema.",
        "description": "Perfil Necessário: Editor ou Admin\r\nExemplo de requisição:\r\n\r\n    POST /Cliente\r\n    {\r\n        \"nome\": \"João Silva\",\r\n        \"email\": \"joao@exemplo.com\",\r\n        \"telefone\": \"11999999999\"\r\n    }",
        "requestBody": {
          "description": "Objeto DTO contendo as informações do cliente",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Retorna o ID do cliente criado",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              }
            }
          },
          "400": {
            "description": "Se os dados do cliente forem inválidos",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Se ocorrer um erro interno durante o processamento"
          }
        }
      },
      "put": {
        "tags": [
          "Cliente"
        ],
        "summary": "Atualiza as informações de um cliente existente.",
        "description": "Perfil Necessário: Editor ou Admin\r\nExemplo de requisição:\r\n\r\n    PUT /Cliente\r\n    {\r\n        \"id\": 1,\r\n        \"nome\": \"João Silva Atualizado\",\r\n        \"email\": \"joao.atual@exemplo.com\",\r\n        \"telefone\": \"11988888888\"\r\n    }",
        "requestBody": {
          "description": "Objeto DTO contendo as informações atualizadas do cliente",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Se o cliente foi atualizado com sucesso",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              }
            }
          },
          "400": {
            "description": "Se os dados atualizados do cliente forem inválidos",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Se o cliente não for encontrado",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Cliente/{clienteId}": {
      "delete": {
        "tags": [
          "Cliente"
        ],
        "summary": "Remove um cliente do sistema.",
        "description": "Perfil Necessário: Editor ou Admin\r\n\r\n    DELETE /Cliente/{clienteId}",
        "parameters": [
          {
            "name": "clienteId",
            "in": "path",
            "description": "O ID do cliente a ser removido",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Se o cliente foi removido com sucesso",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              }
            }
          },
          "404": {
            "description": "Se o cliente não foi encontrado",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Cliente"
        ],
        "summary": "Recupera os detalhes de um cliente específico pelo ID.",
        "description": "Perfil Necessário: Padrão, Editor ou Admin\r\n\r\n    GET /Cliente/{clienteId}",
        "parameters": [
          {
            "name": "clienteId",
            "in": "path",
            "description": "O ID do cliente a ser consultado",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna as informações do cliente solicitado",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              }
            }
          },
          "404": {
            "description": "Se o cliente não foi encontrado",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Cliente/{pagina}/{totalPorPagina}/{termo}": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "summary": "Recupera uma lista paginada de clientes com funcionalidade de busca opcional.",
        "description": "Perfil Necessário: Padrão, Editor ou Admin\r\n\r\n    GET /Cliente/{pagina}/{totalPorPagina}/{termo?}",
        "parameters": [
          {
            "name": "pagina",
            "in": "path",
            "description": "Número da página (padrão: 1)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "totalPorPagina",
            "in": "path",
            "description": "Itens por página (padrão: 10)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "termo",
            "in": "path",
            "description": "Termo de busca opcional para filtrar clientes por nome, email ou telefone",
            "required": true,
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna a lista de clientes",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteListagemDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteListagemDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteListagemDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Se nenhum cliente for encontrado",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ClienteDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "cpf": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "nullable": true
          },
          "endereco": {
            "type": "string",
            "nullable": true
          },
          "observacao": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ClienteListagemDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "cpf": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Token estilo JWT de autorizacao usado no header atraves do esqueema Bearer",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}